[AND-299] Split bundles view and games screen
